+++ examples/CMakeLists.txt
@@ -72,11 +72,11 @@ endif()
 
 if (STDEXEC_ENABLE_TBB)
  add_executable(example.benchmark.tbb_thread_pool benchmark/tbb_thread_pool.cpp)
- target_link_libraries(example.benchmark.tbb_thread_pool PRIVATE STDEXEC::tbbexec)
+ target_link_libraries(example.benchmark.tbb_thread_pool PRIVATE STDEXEC::tbbpool)
 
  add_executable(example.benchmark.tbb_thread_pool_nested benchmark/tbb_thread_pool_nested.cpp)
- target_link_libraries(example.benchmark.tbb_thread_pool_nested PRIVATE STDEXEC::tbbexec)
+ target_link_libraries(example.benchmark.tbb_thread_pool_nested PRIVATE STDEXEC::tbbpool)
 
  add_executable(example.benchmark.fibonacci benchmark/fibonacci.cpp)
- target_link_libraries(example.benchmark.fibonacci PRIVATE STDEXEC::tbbexec)
+ target_link_libraries(example.benchmark.fibonacci PRIVATE STDEXEC::tbbpool)
 endif()
