+++ examples/benchmark/fibonacci.cpp
@@ -17,7 +17,7 @@
 #include <cstdlib>
 #include <iostream>
 
-#include <tbbexec/tbb_thread_pool.hpp>
+#include <execpools/tbb/tbb_thread_pool.hpp>
 #include <exec/static_thread_pool.hpp>
 
 #include <exec/any_sender_of.hpp>
@@ -100,10 +100,10 @@ int main(int argc, char** argv) {
     return -1;
   }
 
-  std::variant<tbbexec::tbb_thread_pool, exec::static_thread_pool> pool;
+  std::variant<execpools::tbb_thread_pool, exec::static_thread_pool> pool;
 
   if (argv[4] == std::string_view("tbb")) {
-    pool.emplace<tbbexec::tbb_thread_pool>(static_cast<int>(std::thread::hardware_concurrency()));
+    pool.emplace<execpools::tbb_thread_pool>(static_cast<int>(std::thread::hardware_concurrency()));
   } else {
     pool.emplace<exec::static_thread_pool>(
       std::thread::hardware_concurrency(), exec::bwos_params{}, exec::get_numa_policy());
