+++ examples/benchmark/tbb_thread_pool.cpp
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 #include "./common.hpp"
-#include <tbbexec/tbb_thread_pool.hpp>
+#include <execpools/tbb/tbb_thread_pool.hpp>
 
 struct RunThread {
   void operator()(
-    tbbexec::tbb_thread_pool& pool,
+    execpools::tbb_thread_pool& pool,
     std::size_t total_scheds,
     std::size_t tid,
     std::barrier<>& barrier,
@@ -85,5 +85,5 @@ struct RunThread {
 };
 
 int main(int argc, char** argv) {
-  my_main<tbbexec::tbb_thread_pool, RunThread>(argc, argv);
+  my_main<execpools::tbb_thread_pool, RunThread>(argc, argv);
 }
