+++ include/execpools/tbb/tbb_thread_pool.hpp
@@ -20,11 +20,11 @@
 #include <tbb/task_arena.h>
 
 #include <exec/static_thread_pool.hpp>
-#include <tbbexec/thread_pool_base.hpp>
+#include <execpools/thread_pool_base.hpp>
 
-namespace tbbexec {
+namespace execpools {
 
-  class tbb_thread_pool : public _thpool::thread_pool_base<tbb_thread_pool> {
+  class tbb_thread_pool : public thread_pool_base<tbb_thread_pool> {
    public:
     //! Constructor forwards to tbb::task_arena constructor:
     template <class... Args>
@@ -44,15 +44,15 @@ namespace tbbexec {
       return stdexec::forward_progress_guarantee::parallel;
     }
 
-    friend _thpool::thread_pool_base<tbb_thread_pool>;
+    friend thread_pool_base<tbb_thread_pool>;
 
     template <class PoolType, class ReceiverId>
-    friend struct _thpool::operation;
+    friend struct operation;
 
-    void enqueue(_thpool::task_base* task, std::uint32_t tid = 0) noexcept {
+    void enqueue(task_base* task, std::uint32_t tid = 0) noexcept {
       arena_.enqueue([task, tid] { task->__execute(task, /*tid=*/tid); });
     }
 
     tbb::task_arena arena_{tbb::task_arena::attach{}};
   };
-} // namespace tbbexec
+} // namespace execpools
