+++ test/execpools/test_tbb_thread_pool.cpp
@@ -26,7 +26,7 @@
 #include <exec/env.hpp>
 #include <exec/inline_scheduler.hpp>
 
-#include <tbbexec/tbb_thread_pool.hpp>
+#include <execpools/tbb/tbb_thread_pool.hpp>
 
 namespace ex = stdexec;
 
@@ -82,9 +82,9 @@ namespace {
   } // namespace
 
   TEST_CASE(
-    "exec::on works when changing threads with tbbexec::tbb_thread_pool",
+    "exec::on works when changing threads with execpools::tbb_thread_pool",
     "[adaptors][exec::on]") {
-    tbbexec::tbb_thread_pool pool;
+    execpools::tbb_thread_pool pool;
     auto pool_sched = pool.get_scheduler();
     CHECK(
       stdexec::get_forward_progress_guarantee(pool_sched)
@@ -104,7 +104,7 @@ namespace {
       return x + 1;
     };
 
-    tbbexec::tbb_thread_pool pool(1);
+    execpools::tbb_thread_pool pool(1);
 
     exec::static_thread_pool other_pool(1);
 
@@ -138,7 +138,7 @@ namespace {
     // state. We'd better have it act normally here.
     using namespace stdexec;
 
-    tbbexec::tbb_thread_pool tbb_pool;
+    execpools::tbb_thread_pool tbb_pool;
     exec::static_thread_pool other_pool(1);
     {
       CHECK_THROWS(stdexec::sync_wait(starts_on(tbb_pool.get_scheduler(), just(0)) | then([](auto) {
@@ -162,7 +162,7 @@ namespace {
   TEST_CASE("tbb_thread_pool async_inclusive_scan") {
     const auto input = std::array{1.0, 2.0, -1.0, -2.0};
     std::remove_const_t<decltype(input)> output;
-    tbbexec::tbb_thread_pool pool{2};
+    execpools::tbb_thread_pool pool{2};
     auto [value] =
       stdexec::sync_wait(async_inclusive_scan(pool.get_scheduler(), input, output, 0.0, 4)).value();
     STATIC_REQUIRE(std::is_same_v<decltype(value), std::span<double>>);
